refactor(mcp/write): #466 extract CLUSTER_C op-dispatch + AST mutation primitives into write_mutation.py (STRATEGY_S1)
Final step (4 of 4) of the v1.13.1 STRATEGY_S1 sequence (parent #459).
Moves the op-dispatch and AST mutation primitives out of write.py into a
new sibling module write_mutation.py, mirroring the pattern established
by predecessors PR #467 (CLUSTER_A → write_detection.py), PR #469
(CLUSTER_B → write_metrics.py), and PR #471 (CLUSTER_D → write_format.py).

Symbols moved verbatim (no signature change, no behaviour change):

- _is_delete_sentinel
- _is_op_descriptor
- _extract_op_descriptor    ($op envelope unwrap)
- _mark_dirty               (Strategy-A dirty-flag bookkeeping; paired-write contract surface)
- _apply_array_op_inplace   ($op APPEND/PREPEND dispatch)
- _target_type_for_assignment
- _normalize_value_for_ast  (JSON→AST node coercion; v1.14.0 Case A #460 home)

Co-migrations (helpers exclusively consumed by the moved set, mirroring
PR #471's FORMAT_STYLE co-migration discipline):

- _KNOWN_OPS frozenset (only used by _extract_op_descriptor)

Importer rewiring:

- src/octave_mcp/mcp/write.py: now imports the 7 symbols from
  write_mutation; replaces the in-place definitions with a comment
  documenting the extraction and the v1.14.0 #460 Case A home.
- src/octave_mcp/cli/main.py: _mark_dirty import switched to
  write_mutation; import block re-sorted for ruff I001.
- tests/unit/test_literal_zones_write.py: _normalize_value_for_ast
  import switched to write_mutation.
- tests/unit/test_write_tool.py: _normalize_value_for_ast import
  switched to write_mutation.

No re-export shims (parent issue acceptance discipline).
Also removes the now-unused LiteralZoneValue import from write.py.

LOC delta:
- write.py: 3407 → 3197 (-210 LOC)
- write_mutation.py: new, 240 LOC
- Cumulative STRATEGY_S1 reduction: 4887 → 3197 (-1690 LOC across CLUSTER A/B/D/C)

Quality gates (all blocking, all green):
- tests/unit/test_dirty_paired_write_lint.py: 3/3 passed
  (the paired-write contract fence; proximity-based detection of
  _mark_dirty within ±10 lines of every node.value mutation survived
  the module boundary byte-identically — all _mark_dirty call sites
  remain inline at their mutation locations, only the function
  definition moved)
- Full pytest suite: 3614 passed, 11 skipped, 3 xfailed
  (zero test count regression vs. baseline; pre-existing baseline
  count incorporates predecessor PR #467/#469/#471 test additions)
- ruff/black/mypy: clean

Completes STRATEGY_S1. write.py is now at its parent-issue target and
ready for v1.14.0 hybrid fix work (#460 Case A literal-zone form
preservation) which will modify _normalize_value_for_ast in its new
write_mutation.py home.

Refs: #466, #459. Predecessors: #467, #469, #471.
